Refatorando Front-end e validando as querys

+++ public/src/Route/category-delete.php
@@ -5,11 +5,15 @@ require_once '../Helpers/user-session.php';
 
 try {
 	$category = new CategoryDao($_POST['name'], $_POST['id']);
-	if ($category->delete()) {
-		$_SESSION['success'] = "<span>{$_POST['name']}</span> removido com sucesso";
+	if (is_numeric($_POST['id']) && $category->verifyCategoryExist()) {
+		if ($category->delete()) {
+			$_SESSION['success'] = "<span>{$_POST['name']}</span> removido com sucesso";
+			header("Location: ../View/category.php");
+		}		
+	} else {
+		$_SESSION['danger'] = "<span>{$_POST['name']}</span> não foi removido, ID  <span>{$_POST['id']}</span> inválido";
 		header("Location: ../View/category.php");
 	}
-	die();	
 } catch (PDOException $e) {
 	Erro::handler($e);
 	$_SESSION['danger'] = "<span>{$_POST['name']}</span> não foi removido";
